+++ ruma-client-api/CHANGELOG.md
@@ -21,6 +21,15 @@ Breaking changes:
 * Update type of `user_id` in `r0::account::whoami` from `String` to `ruma_identifiers::UserId`
 * Update type of `limited` in `r0::sync::sync_events::Timeline` from `Option<bool>` to `bool`
 * Use `DeviceId` for `device_id` field of `r0::session::login::Response`
+* Use `ruma_identifiers::ServerName` instead of `String` for `server_name` fields in the following endpoints:
+  ```rust
+  r0::{
+      account::request_openid_token,
+      media::{get_content, get_content_as_filename, get_content_thumbnail},
+      membership::join_room_by_id_or_alias,
+      session::login,
+  }
+  ```
 
 Improvements: